requirements.txt
asgiref==3.9.2
attrs==25.3.0
Django==5.2.6
django-cuser==2017.3.16
django-filter==25.1
django-jazzmin==3.0.1
django-modeltranslation==0.19.7
django-tinymce==4.1.0
django-versatileimagefield==3.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-magic==0.4.27
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
uritemplate==4.2.0
